+++ cookbooks/NodeJs/metadata.rb
@@ -1,7 +1,7 @@
-name             'angular'
+name             'NodeJs'
 maintainer       'NA'
 maintainer_email 'eochoa260@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures angular'
+description      'Installs/Configures NodeJs'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'
+version          '0.2.0'