+++ cookbooks/NodeJs/recipes/config.rb
@@ -1,0 +1,5 @@
+template "#{node['NodeJs']['home']['dir']}/server.js" do 
+	source 'server.js.erb'
+	mode '0755'
+end
+