+++ cookbooks/angular/recipes/config.rb
@@ -1,5 +1,0 @@
-template '/opt/angular/server.js' do 
-	source 'server.js.erb'
-	mode '0755'
-end
-