+++ cookbooks/angular/recipes/directory.rb
@@ -1,9 +1,0 @@
-directory '/opt/angular' do
-	action :create
-	mode '0775'
-end
-
-directory '/opt/angular/angularjs' do
-	action :create
-	mode '0755'
-end