+++ cookbooks/todo/metadata.rb
@@ -6,4 +6,4 @@ description      'Installs/Configures todo'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends "angular", "~> 0.1.0"
+depends "NodeJs", "~> 0.1.0"